Add Tilemill setup and project
To run Tilemill just run 'tilemill/tilemill'.

Tilemill is used to create the map style. It is not setup to use ALL the
map resources and instead only uses a sample area. This allows for a
nicer development experience as the whole map will not have to be
rendered. Currently the tilemill/map_source are downloaded from static
assets hosted on AWS S3. Eventually these assets should be moved to the
root map_source directory along with a file to download and process just
the Tilemill sources.